fix(gym_env): remove float64->float32 warning for value +-1.0
The implicit dtype for a gym.spaces.Box is float32. The default dtype for an np.ndarray is float64. The boxes here are created with -1.0 or 1.0 values exclusively. Thus, precision is not an issue here.

Make the dtype for the np.ndarrays given to the Box c'tors explicitly float32.

+++ src/python/double_pendulum/simulation/gym_env.py
@@ -12,9 +12,13 @@ class CustomEnv(gym.Env):
         terminated_func,
         reset_func,
         obs_space=gym.spaces.Box(
-            np.array([-1.0, -1.0, -1.0, -1.0]), np.array([1.0, 1.0, 1.0, 1.0])
+            np.array([-1.0, -1.0, -1.0, -1.0], dtype=np.float32),
+            np.array([1.0, 1.0, 1.0, 1.0], dtype=np.float32),
         ),
-        act_space=gym.spaces.Box(np.array([-1.0, -1.0]), np.array([1.0, 1.0])),
+        act_space=gym.spaces.Box(
+            np.array([-1.0, -1.0], dtype=np.float32),
+            np.array([1.0, 1.0], dtype=np.float32),
+        ),
         max_episode_steps=1000,
         scaling=True,
     ):
